requirements.txt
# Pipecat AI with required features
pipecat-ai[websocket,openai,silero,sarvam,google]>=0.0.86

# Vobiz telephony integration (from PyPI)
pipecat-vobiz>=0.0.2

# Web framework
fastapi
uvicorn
python-multipart

# HTTP client
aiohttp

# Environment variables
python-dotenv

# Supabase
supabase

# Logging
loguru
